requirements.txt
dspy-ai==2.6.27
litellm
rich
python-dotenv
